--- base/uk.ac.stfc.isis.ibex.ui.graphing/src/uk/ac/stfc/isis/ibex/ui/graphing/websocketview/PlotUpdateAdapter.java
@@ -1,26 +1,0 @@
-package uk.ac.stfc.isis.ibex.ui.graphing.websocketview;
-
-/**
- * Adapter for the IPlotUpdateListener interface.
- */
-public class PlotUpdateAdapter implements IPlotUpdateListener {
-
-	/**
-	 * {@inheritDoc}
-	 */
-	@Override
-	public void imageUpdated() {}
-	
-	/**
-	 * {@inheritDoc}
-	 */
-	@Override
-	public void onConnectionStatus(boolean isConnected) {}
-
-	/**
-	 * {@inheritDoc}
-	 */
-	@Override
-	public void onPlotNameChange(String newName) {}
-    
-}
